clean up after ourselves and remove the old QB code and route handling.
