+++ frontend/src/metabase/app.js
@@ -21,7 +21,6 @@ import "./services";
 import "./icons";
 
 import "./auth/auth.module";
-import "./card/card.module";
 import "./dashboard/dashboard.module";
 import "./home/home.module";
 import "./pulse/pulse.module";
@@ -58,7 +57,6 @@ var Metabase = angular.module('metabase', [
     'metabase.directives',
     'metabase.controllers',
     'metabase.icons',
-    'metabase.card',
     'metabase.dashboard',
     'metabase.home',
     'metabase.pulse',
@@ -150,6 +148,24 @@ Metabase.config(['$routeProvider', '$locationProvider', function($routeProvider,
     $routeProvider.when('/admin/', {
         redirectTo: function(routeParams, path, search) {
             return '/admin/settings';
+        }
+    });
+
+    // redirect old urls to new ones with hashes
+    $routeProvider.when('/q/:serializedCard', {
+        redirectTo: function (routeParams, path, search) {
+            return "/q#"+routeParams.serializedCard;
+        }
+    });
+    $routeProvider.when('/card/:cardId/:serializedCard', {
+        redirectTo: function (routeParams, path, search) {
+            return "/card/"+routeParams.cardId+"#"+routeParams.serializedCard;
+        }
+    });
+
+    $routeProvider.when('/card/', {
+        redirectTo: function (routeParams, path, search) {
+            return "/questions/all";
         }
     });
